% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univ.R
\name{univariate}
\alias{univariate}
\title{ULM with Significance}
\usage{
univariate(data, out, alpha)
}
\arguments{
\item{data}{is dataset}

\item{out}{is response variable}

\item{alpha}{is alpha level for significance test}
}
\value{
dataframe of predictors with p-values and
significance, smallest p-value (most significant), and
scatterplot of response variable and most significant
predictor in univariate linear regression model
}
\description{
Description
}
\details{
Perform Univariate Linear Regression on a dataset
}
\author{
Emily King
}
